Rexture more of the cockpit
Add canvas PFD. The SVG / Nasal only displays two parameters currently; but the menu selection and page framework is all wired in.

There will need to be multiple instances of this for each display. I think each display is capable of the same functions ; but not sure about that.

From: Richard Harrison <rjh@zaretto.com>

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/SpaceShuttle@612 9eb9bc84-db9d-4e05-893b-26dd41ac94a6