Add canvas HUD for pilot / commander.
These are based on the F-15 and need to be adapted more for the shuttle; however it is a reasonable starting point.

The geometry / position is about right (in the 3d model).

Next steps are to make the SVG more like the real thing (2.7-22) and add support for on off switch and the declutter modes.